refactor: clarify crm adapter boundary
