+++ apps/web/src/adapters/crm/forecast-snapshot-observability.test.ts
