+++ apps/web/src/adapters/crm/pipeline-repository.ts
