+++ apps/web/src/app/[locale]/(staff)/staff/crm/_core.ts
@@ -12,7 +12,7 @@ import {
 } from '@interdomestik/domain-crm/reporting';
 import type { CrmReportingRepository } from '@interdomestik/domain-crm/reporting/repository';
 
-import { crmReportingRepository } from '@/lib/domain-crm/reporting-repository';
+import { crmReportingRepository } from '@/adapters/crm/reporting-repository';
 
 export const STAFF_CRM_REPORTING_WINDOW_DAYS = 90;
 export const STAFF_CRM_PIPELINE_MAX_ROWS = 10;
