+++ apps/web/src/lib/actions/agent/update-lead-status.core.ts
@@ -3,7 +3,7 @@ import {
   type CrmLeadMutationResult,
 } from '@interdomestik/domain-crm/leads/mutations';
 import type { CrmActorContext } from '@interdomestik/domain-crm/context';
-import { crmLeadMutationRepository } from '@/lib/domain-crm/lead-mutation-repository';
+import { crmLeadMutationRepository } from '@/adapters/crm/lead-mutation-repository';
 
 function mapStageError(result: Extract<CrmLeadMutationResult, { success: false }>) {
   if (result.error === 'not_found') {
