+++ package.json
@@ -56,6 +56,7 @@
     "mod:autopilot:apply": "node scripts/modularize-autopilot.mjs --apply",
     "check:entrypoints": "node scripts/check-entrypoints-no-db.mjs",
     "check:db-access": "node scripts/check-db-access-guard.mjs",
+    "check:architecture-boundaries": "node scripts/check-architecture-boundaries.mjs",
     "check:evidence-storage-paths": "node scripts/check-evidence-storage-paths.mjs",
     "check:entrypoints:strict": "node scripts/check-entrypoints-no-db.mjs --strict",
     "test:release-gate": "tsx --test scripts/release-gate/run.test.ts scripts/release-gate/pilot-artifacts.test.ts scripts/release-gate/pilot-evidence-workflow.test.ts",
@@ -81,7 +82,7 @@
     "e2e:gate:mk:fast": "E2E_DATABASE_URL=${E2E_DATABASE_URL:-postgresql://postgres:postgres@127.0.0.1:54322/postgres} E2E_DATABASE_URL_RLS=${E2E_DATABASE_URL_RLS:-postgresql://postgres:postgres@127.0.0.1:54322/postgres} NEXT_PUBLIC_BILLING_TEST_MODE=1 ./scripts/m4-gatekeeper.sh && E2E_DATABASE_URL=${E2E_DATABASE_URL:-postgresql://postgres:postgres@127.0.0.1:54322/postgres} E2E_DATABASE_URL_RLS=${E2E_DATABASE_URL_RLS:-postgresql://postgres:postgres@127.0.0.1:54322/postgres} NEXT_PUBLIC_BILLING_TEST_MODE=1 PW_FAST_GATES=1 pnpm --filter @interdomestik/web exec playwright test e2e/gate --project=gate-mk-mk --max-failures=1 --trace=retain-on-failure --reporter=line",
     "check:fast": "node scripts/run-with-default-db-url.mjs bash -c 'cd \"$PWD\" && NEXT_PUBLIC_BILLING_TEST_MODE=1 pnpm --filter @interdomestik/web run build:ci' && node scripts/run-with-default-db-url.mjs pnpm e2e:state:setup && node scripts/run-with-default-db-url.mjs pnpm e2e:gate:pr:fast",
     "check:static": "pnpm repo:size:check && pnpm format:check && pnpm lint && pnpm type-check",
-    "check:all": "pnpm repo:size:check && pnpm format:check && pnpm lint && pnpm type-check && pnpm i18n:check && pnpm i18n:purity:check && pnpm check:entrypoints:strict && pnpm check:db-access && pnpm test:unit:domains && pnpm test",
+    "check:all": "pnpm repo:size:check && pnpm format:check && pnpm lint && pnpm type-check && pnpm i18n:check && pnpm i18n:purity:check && pnpm check:entrypoints:strict && pnpm check:db-access && pnpm check:architecture-boundaries && pnpm test:unit:domains && pnpm test",
     "check": "pnpm check:fast && pnpm build",
     "ci:local:quick": "bash scripts/ci-local-parity.sh quick",
     "ci:local:pr": "bash scripts/ci-local-parity.sh pr",
@@ -148,7 +149,7 @@
     "v1:status": "node scripts/v1-status.mjs",
     "ci:audit:docker": "bash scripts/ci-audit-docker.sh",
     "purity:audit": "node scripts/purity-audit.mjs",
-    "pr:verify": "pnpm memory:precheck && pnpm repo:size:check && pnpm test:release-gate && pnpm check:e2e-contracts && pnpm lint:production-warnings && pnpm db:migrations:check-journal && pnpm db:rls:test:required && pnpm i18n:check && pnpm i18n:purity:check && pnpm check:db-access && pnpm coverage:gate && pnpm check:fast && node scripts/run-with-default-db-url.mjs pnpm --filter @interdomestik/web run e2e:smoke",
+    "pr:verify": "pnpm memory:precheck && pnpm repo:size:check && pnpm test:release-gate && pnpm check:e2e-contracts && pnpm lint:production-warnings && pnpm db:migrations:check-journal && pnpm db:rls:test:required && pnpm i18n:check && pnpm i18n:purity:check && pnpm check:db-access && pnpm check:architecture-boundaries && pnpm coverage:gate && pnpm check:fast && node scripts/run-with-default-db-url.mjs pnpm --filter @interdomestik/web run e2e:smoke",
     "pr:verify:hosts": "bash scripts/pr-verify-hosts.sh",
     "pr:finalize": "bash scripts/pr-finalizer.sh",
     "multiagent:preflight": "bash scripts/multi-agent/preflight-agent.sh",
